app/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.5
dnspython==2.8.0
dotenv==0.9.9
email-validator==2.3.0
fastapi==0.129.0
greenlet==3.3.1
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
pwdlib==0.3.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.0
pydantic_core==2.41.5
PyJWT==2.11.0
python-dotenv==1.2.1
python-multipart==0.0.22
SQLAlchemy==2.0.46
sqlmodel==0.0.33
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0